incorporated rest controller, added unit test and fix type mismatch for repo result with dtos

+++ src/main/java/org/test/wsd/testcommerce/dto/ItemSaleAmountDto.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @Builder
 public class ItemSaleAmountDto {
     private Long id;
-    private int totalSaleAmount;
+    private Long totalSaleAmount;
 }
